LayoutTests/fast/canvas/webgl/tex-image-and-sub-image-2d-with-image.html

<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css"/>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
<script src="resources/utils3d.js"> </script>
<script id="vshader" type="x-shader/x-vertex">
attribute vec3 g_Position;
attribute vec2 g_TexCoord0;

varying vec2 texCoord;

void main()
{
    gl_Position = vec4(g_Position.x, g_Position.y, g_Position.z, 1.0);
    texCoord = g_TexCoord0;
}
</script>

<script id="fshader" type="x-shader/x-fragment">
uniform sampler2D tex;
varying vec2 texCoord;

void main()
{
    gl_FragColor = texture2D(tex, texCoord);
}
</script>

<script>
var gl = null;
var textureLoc = null;
var successfullyParsed = false;

function init()
{
    if (window.layoutTestController) {
        layoutTestController.overridePreference("WebKitWebGLEnabled", "1");
        layoutTestController.dumpAsText();
        layoutTestController.waitUntilDone();
    }

    description('Verify texImage2D and texSubImage2D code paths taking Images, in particular flipY argument');

    debug('Regression test for <a href="https://bugs.webkit.org/show_bug.cgi?id=34458">https://bugs.webkit.org/show_bug.cgi?id=34458</a> : <code>Refactor texImage2D and texSubImage2D taking Image to use common code</code>');

    gl = initWebGL("example", "vshader", "fshader", [ "g_Position", "g_TexCoord0" ], [ 0, 0, 0, 1 ], 1);

    textureLoc = gl.getUniformLocation(gl.program, "tex");

    var vertices = new WebGLFloatArray([
         1.0,  1.0, 0.0,
        -1.0,  1.0, 0.0,
        -1.0, -1.0, 0.0,
         1.0,  1.0, 0.0,
        -1.0, -1.0, 0.0,
         1.0, -1.0, 0.0]);
    var texCoords = new WebGLFloatArray([
        1.0, 1.0,
        0.0, 1.0,
        0.0, 0.0,
        1.0, 1.0,
        0.0, 0.0,
        1.0, 0.0]);
    var texCoordOffset = vertices.byteLength;

    var vbo = gl.createBuffer();
    gl.bindBuffer(gl.ARRAY_BUFFER, vbo);
    gl.bufferData(gl.ARRAY_BUFFER,
                  texCoordOffset + texCoords.byteLength,
                  gl.STATIC_DRAW);
    gl.bufferSubData(gl.ARRAY_BUFFER, 0, vertices);
    gl.bufferSubData(gl.ARRAY_BUFFER, texCoordOffset, texCoords);

    gl.enableVertexAttribArray(0);
    gl.vertexAttribPointer(0, 3, gl.FLOAT, false, 0, 0);
    gl.enableVertexAttribArray(1);
    gl.vertexAttribPointer(1, 2, gl.FLOAT, gl.FALSE, 0, texCoordOffset);

    loadTexture("resources/red-green.png");
}

function loadTexture(src) {
    var texture = gl.createTexture();
    gl.bindTexture(gl.TEXTURE_2D, texture);
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
    var image = new Image();
    image.onload = function() {
        gl.bindTexture(gl.TEXTURE_2D, texture);
        gl.texImage2D(gl.TEXTURE_2D, 0, image, true);
        runTest(image);
    };
    image.src = src;
    return texture;
}

// These two declarations need to be global for "shouldBe" to see them
var buf = null;
var idx = 0;
var pixel = [0, 0, 0];
var correctColor = null;

function runOneIteration(image, useTexSubImage2D, flipY, topColor, bottomColor)
{
    debug('Testing ' + (useTexSubImage2D ? 'texSubImage2D' : 'texImage2D') +
          ' with flipY=' + flipY);
    gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
    // Disable any writes to the alpha channel
    gl.colorMask(1, 1, 1, 0);
    var texture = gl.createTexture();
    // Bind the texture to texture unit 0
    gl.bindTexture(gl.TEXTURE_2D, texture);
    // Set up texture parameters
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
    // Upload the image into the texture
    if (useTexSubImage2D) {
        // Initialize the texture to black first
        gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGB, image.width, image.height, 0,
                      gl.RGBA, gl.UNSIGNED_BYTE, null);
        gl.texSubImage2D(gl.TEXTURE_2D, 0, 0, 0, image, flipY, false);
    } else {
        gl.texImage2D(gl.TEXTURE_2D, 0, image, flipY, false);
    }

    // Point the uniform sampler to texture unit 0
    gl.uniform1i(textureLoc, 0);
    // Draw the triangles
    gl.drawArrays(gl.TRIANGLES, 0, 6);

    // Read back the rendering results
    var width = 32;
    var height = 32;
    buf = gl.readPixels(0, 0, width, height, gl.RGBA, gl.UNSIGNED_BYTE);
    // Check a few pixels near the top and bottom and make sure they have
    // the right color.
    var queryWidth = 2;
    var queryHeight = 2;
    debug("Checking lower left corner");
    var xoff = 4;
    var yoff = 4;
    correctColor = bottomColor;
    for (var y = 0; y < queryHeight; y++) {
        for (var x = 0; x < queryWidth; x++) {
            idx = ((yoff + y) * width * 4 +
                   (xoff + x) * 4);
            pixel[0] = buf[idx];
            pixel[1] = buf[idx + 1];
            pixel[2] = buf[idx + 2];
            shouldBe("pixel", "correctColor");
        }
    }
    debug("Checking upper left corner");
    yoff = height - 8;
    correctColor = topColor;
    for (var y = 0; y < queryHeight; y++) {
        for (var x = 0; x < queryWidth; x++) {
            idx = ((yoff + y) * width * 4 +
                   (xoff + x) * 4);
            pixel[0] = buf[idx];
            pixel[1] = buf[idx + 1];
            pixel[2] = buf[idx + 2];
            shouldBe("pixel", "correctColor");
        }
    }
}

function runTest(image)
{
    var red = [255, 0, 0];
    var green = [0, 255, 0];
    runOneIteration(image, false, true, red, green);
    runOneIteration(image, false, false, green, red);
    runOneIteration(image, true, true, red, green);
    runOneIteration(image, true, false, green, red);

    successfullyParsed = true;
    var epilogue = document.createElement("script");
    epilogue.onload = finish;
    epilogue.src = "../../js/resources/js-test-post.js";
    document.body.appendChild(epilogue);
}

function finish() {
    if (window.layoutTestController) {
        layoutTestController.notifyDone();
    }
}
</script>
</head>
<body onload="init()">
<canvas id="example" width="32px" height="32px"></canvas>
<div id="description"></div>
<div id="console"></div>
</body>
</html>
