LayoutTests/fast/canvas/webgl/tex-image-and-sub-image-2d-with-image-expected.txt

Verify texImage2D and texSubImage2D code paths taking Images

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing texImage2D with flipY=true
Checking lower left corner
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
Checking upper left corner
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
Testing texImage2D with flipY=false
Checking lower left corner
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
Checking upper left corner
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
Testing texSubImage2D with flipY=true
Checking lower left corner
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
Checking upper left corner
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
Testing texSubImage2D with flipY=false
Checking lower left corner
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
Checking upper left corner
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS pixel is correctColor
PASS successfullyParsed is true

TEST COMPLETE
